[refactor] redux designer context 수정
