+++ client/src/Contexts/Designer/StyleScrap/action.js
@@ -1,6 +1,6 @@
 import { REGISTER_STYLESCRAP, GET_STYLESCRAP_LIST, DELETE_STYLESCRAP } from './constant';
-import StyleScrapAPI from '../../Api/styleScrap';
-import { createPromiseThunk } from '../Common/asyncUtils';
+import StyleScrapAPI from '../../../Api/styleScrap';
+import { createPromiseThunk } from '../../Common/asyncUtils';
 
 export const registerStyleScrap = (payload) => ({ type: REGISTER_STYLESCRAP, payload });
 export const deleteStyleScrap = (id) => ({ type: DELETE_STYLESCRAP, id });
